requirements.txt
Django~=4.2.3
djangorestframework~=3.14.0
python-decouple~=3.8
PySocks~=1.7.1
django-cors-headers~=4.2.0
requests~=2.31.0
gunicorn~=21.2.0